+++ .github/workflows/js-ci.yaml
@@ -63,4 +63,5 @@ jobs:
           cache-dependency-path: js/yarn.lock
       - run: corepack enable
       - run: yarn --immutable
+      - run: yarn workspace @pluralsh/design-system build
       - run: yarn test
