+++ js/console/src/components/workbenches/workbench/job/WorkbenchJobActivityResults.tsx
@@ -37,7 +37,7 @@ import {
   WorkbenchJobActivityTraceFragment,
   WorkbenchToolQueryData,
 } from 'generated/graphql'
-import { groupBy, isEmpty, isNil } from 'lodash'
+import { isEmpty } from 'lodash'
 import {
   ComponentPropsWithRef,
   ComponentType,
@@ -52,6 +52,7 @@ import { COLORS } from 'utils/color'
 import { formatDateTime, toDateOrUndef } from 'utils/datetime'
 import { isNonNullable } from 'utils/isNonNullable'
 import { getOldContentFromTextDiff } from 'utils/textDiff'
+import { getMetricSeries, type MetricSeries } from './workbenchJobMetrics'
 import { TraceWaterfall } from './WorkbenchJobTraces'
 
 export function MemoActivityIcon({
@@ -637,55 +638,6 @@ export function WorkbenchJobMetricsLegend({
         </Flex>
       ))}
     </Flex>
-  )
-}
-
-type MetricSeries = {
-  data: { x: Date; y: number }[]
-  id: string
-  label: string
-}
-
-export function getMetricSeries(
-  metrics: WorkbenchJobActivityMetricFragment[]
-): MetricSeries[] {
-  const grouped = groupBy(metrics, metricSeriesId)
-
-  return Object.entries(grouped).map(([id, points]) => ({
-    id,
-    label: metricSeriesLabel(points[0]),
-    data: points
-      .map((point) => ({ x: toDateOrUndef(point.timestamp), y: point.value }))
-      .filter(
-        (point): point is { x: Date; y: number } =>
-          !isNil(point.x) && !isNil(point.y)
-      ),
-  }))
-}
-
-function metricSeriesId({
-  name,
-  labels,
-}: WorkbenchJobActivityMetricFragment): string {
-  return `${name ?? 'metric'}{${metricLabelEntries(labels)
-    .map(([key, value]) => `${key}:${value}`)
-    .join(',')}}`
-}
-
-function metricSeriesLabel({
-  name,
-  labels,
-}: WorkbenchJobActivityMetricFragment): string {
-  const label = metricLabelEntries(labels)
-    .map(([key, value]) => `${key}=${value}`)
-    .join(', ')
-
-  return label || name || 'metric'
-}
-
-function metricLabelEntries(labels: Nullable<Record<string, unknown>>) {
-  return Object.entries(labels ?? {}).sort(([left], [right]) =>
-    left.localeCompare(right)
   )
 }
